Move common functionality to ExactTargetApiTestBase
- move setup to the base classe
- move the logger creator
- move the soap client creator
